+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## v1.1 (13/04/2023)
 
+### fixes
+
+ - [fix: #19] - ensure actions value is an array
+ 
+### enhancements
+
  - add custom repository to each component
  - add Grid, View and Message component
  - add docker configurations